[#noissue] Extract OtlpAttributeBoMapper bean and reduce TransformContext to TruncationCounter
